+++ src/test/compile-fail/deprecated-auto-code.rs
@@ -1,0 +1,15 @@
+// Copyright 2013 The Rust Project Developers. See the COPYRIGHT
+// file at the top-level directory of this distribution and at
+// http://rust-lang.org/COPYRIGHT.
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+#[auto_encode] //~ ERROR: `#[auto_encode]` is deprecated
+#[auto_decode] //~ ERROR: `#[auto_decode]` is deprecated
+struct A;
+
+fn main() {}
